+++ pcap4j-sample/src/main/java/org/pcap4j/sample/PacketStream.java
@@ -1,0 +1,66 @@
+package org.pcap4j.sample;
+
+import org.pcap4j.core.NotOpenException;
+import org.pcap4j.core.PcapHandle;
+import org.pcap4j.core.PcapNativeException;
+import org.pcap4j.core.PcapNetworkInterface;
+import org.pcap4j.core.PcapNetworkInterface.PromiscuousMode;
+import org.pcap4j.core.PcapPacket;
+import org.pcap4j.util.NifSelector;
+import org.pcap4j.util.Packets;
+
+import java.io.IOException;
+import java.util.stream.Stream;
+
+@SuppressWarnings("javadoc")
+public class PacketStream {
+
+  private static final String COUNT_KEY
+    = PacketStream.class.getName() + ".count";
+  private static final int COUNT
+    = Integer.getInteger(COUNT_KEY, 5);
+
+  private static final String READ_TIMEOUT_KEY
+    = PacketStream.class.getName() + ".readTimeout";
+  private static final int READ_TIMEOUT
+    = Integer.getInteger(READ_TIMEOUT_KEY, 10); // [ms]
+
+  private static final String SNAPLEN_KEY
+    = PacketStream.class.getName() + ".snaplen";
+  private static final int SNAPLEN
+    = Integer.getInteger(SNAPLEN_KEY, 65536); // [bytes]
+
+  private PacketStream() {}
+
+  public static void main(String[] args) throws PcapNativeException, NotOpenException {
+    System.out.println(COUNT_KEY + ": " + COUNT);
+    System.out.println(READ_TIMEOUT_KEY + ": " + READ_TIMEOUT);
+    System.out.println(SNAPLEN_KEY + ": " + SNAPLEN);
+    System.out.println("\n");
+
+    PcapNetworkInterface nif;
+    try {
+      nif = new NifSelector().selectNetworkInterface();
+    } catch (IOException e) {
+      e.printStackTrace();
+      return;
+    }
+
+    if (nif == null) {
+      return;
+    }
+
+    System.out.println(nif.getName() + "(" + nif.getDescription() + ")");
+
+    try (
+      PcapHandle handle = nif.openLive(SNAPLEN, PromiscuousMode.PROMISCUOUS, READ_TIMEOUT);
+      Stream<PcapPacket> stream = handle.stream()
+    ) {
+      stream
+        .limit(COUNT)
+        .filter(Packets::containsUdpPacket)
+        .forEach(System.out::println);
+    }
+  }
+
+}